review/modifications to swe_bench (#560)
* review/modifications to swe_bench

- Break scoring into its own module to make the main task file a bit more comprehensible.

- Be careful to only import swebench and docker when used (this avoids making these implicit requirements for the package, as the _registry.py file will be executed when any inspect_eval is used)

- Rename scoring functions to be consistent w/ swe_bench (e.g. swe_bench_scorer)

- Rename build_images_for_samples to build_images (just a bit more pithy)

- Move the default solver into a function so it isn't created on import

- Move monkeying w/ global loggers into task function so we only do it when using swe_bench (minimise blast radius in case this has unexpected interactions)

* add core inspect_evals dependencies to swe_bench

---------
