+++ pyproject.toml
@@ -111,7 +111,7 @@ inspect_evals = [
     "datasets",
     "pillow"
 ]
-swe_bench = ["swebench","docker"]
+swe_bench = ["inspect_ai[inspect_evals]", "swebench","docker"]
 
 dev = [
     "anthropic",
